Keyword store and retreive working
A merchant node sets a keyword in the dht using hash(keyword) as the key and
and a dictionary of {contract ID : node info} as the value. 'Node info'
is a serialized kprotocol.Node object containing info about the node that is
needed to contact the merchant store as well as add the node to the routing
table.

TODO:
The store module needs to be made to accept dictionaries as values.

If a value exists the dictionary should be extended with the new value.

The ttl cull needs to check each dictionary to remove expired entries.

We also need a extension of IStorage that uses sqlite instead of memory.

Probably need to extend the kprotocol.Node protobuf to specify a separate
port for the merchant server.