Upgraded ForgetfulStorage to work with dictionary values
The ForgetfulStorage class now stores data as:
{keyword : {Contract ID : Serialized kprotocol.Node}}

When retrieving data out of storage it returns a list of serialized 'Value' protobuf
objects which contain both the contract ID and the node object.

Also the iteritems method has been updated to reflect the change and a new
iterkeys method has been added. Both iteritems and iterkeys are only used by
KademliaProtocol.transferKeyValues when sharing values with new neighbors.

TODO:

The only thing left to do is fix the cull method in ForgetfulStorage which
removes items older than the ttl. Since we now store values as dictionaries
 it needs to sort through the values of each dictionary to find items older
than the ttl. This means iteritemsOlderThan and _tripleIterable need to be
fixed.