ForgetfulStorage fully functional
Added a TTLDict class to Storage to make it easier to delete expired data
from storage.

Data are now stored as a dictionary of {keyword : TTLDict()}.

It appears to be working properly. I will need to write some unit tests and
do more testing.

Also, up to this point the ForgetfulStorage class has been storing data in
memory. This may or may not be useful for OpenBazaar. Certainly we want some
longer lived nodes (merchants?) to store the data on disk. For that we will
use SQLite. Storing data in memory might be useful for short lived nodes, such
as people who are just browsing stores.