Bug fix in crawling
_handleFoundValues in crawling uses collections.Counter to make sure no
conflicting values were returned.

However, we were passing in a list of lists which Counter doesn't accept
because lists are mutable.

Changed RPCProtocol.datagramReceived to return data as a tuple instead of a
list. That fixes the bug.

It appears the library is fully functional now. Will do more testing.