Store own store request on node if key falls within our bucket
Fixed a small bug which wasn't storing our own store requests on our node
if it was in our bucket. The bug caused misbehavior when transferring
key/values to new nodes.

TODO:

There is still a bug in KademliaProtocol.transferKeyValues. Lines 109 and 110
handle the logic of determining when to send a store request to a new node,
but they are not returning the correct values.

When a node (which should be storing values) joins the network, existing nodes
fail to transfer key/values because these two lines.