transferKeyValues fixed
It appears that transferKeyValues is now sending stored values to the correct
nodes. There may be a problem with redundant store requests refreshing the ttl
of stored items. When a new node enters the network it typically shares its
values with nodes as it learns about them, resetting the ttl.

It may be a good idea to prevent ttl refreshing and require new store messages
after expiration.