Additional Node fields sent over the wire.
The additional node fields that were added in the previous commit can now be
sent over the wire. Nodes will store that data with each node object in their
routing tables.

This will allow use to filter our contacts by type (buyer, seller, notary,
etc).

It appears to be working 100% but this change has the potential to introduce
bugs in a number of places so some testing is needed.