Added a delete message to the protocol
Added a new message to the wire protocol for deleting data that we have stored
on other nodes. Since we don't want just anyone to be able to delete our data
we require a signature (coving the 'key' value) be submitted with the delete
request. Since the data stored on each node is a serialized kprotocol.Node
object, it has a field where we can place our public key when making a store
request. If that key is present the node will verify the submitted signature
against it. If we failed, to include our public key when we stored the dated,
the delete message will be ignored and we will just have to wait for expiration.