Added sqlite datastore
Added a PersistentStorage class which implements IStorage. This class is
incomplete at the moment. Both the iterkeys and iteritems methods in the
interface return dictionary iterators which are used by the transferKeyValues
method in KademliaProtocol.

We're going to have to refactor those methods since it isn't appropriate to
return iterators when using sqlite.