Added C module for generating the GUID
To prevent attacks on the DHT we generate the GUID using sha512(signed public
key) where the first 20 bytes are the GUID and the last 32 bytes are a 'proof
of work'.  The PoW must be below a certain treshold for the GUID to be valid.
The target threshold is set to require about 30 seconds of brute forcing, on
average, to find a vaild GUID.

Unfortunately, python is very slow and an attacker could optimize the GUID
creation by using a faster language.

To counter, we generate the GUID in C using a python extension module.