txrudp migration complete
The txrudp migration is just about complete. The only other thing is to do
is take a look at the timeout handling. The dht lib currently handles timeouts
by itself, but the txrudp also has timeout handling. It may be more efficent
to use the txrudp timeout but that requires a number of changes to dht.rcpudp.