Added a check for valid guid
Added a check on the guid to make sure it's valid. This can still be optimized.
We only need to send the signed pubkey/verify the guid parameters on first
contact. After that we don't need to send/verify it again.

That will require some changes to how we determine if a node is new or not.