PING responds with serialized node info
The PING command now responds with the serialized node information. This is
useful because during bootstrapping we were adding nodes to the routing table
based on the response to the PING, giving an attacker an opportunity to slip
us fake GUIDs. Now we validate the GUID in reponse to the ping so no unvalidated
nodes enter our routing table.