Added httpseed/network crawler for seeding
The server crawls the network and maintains a list of just about every node in
the network. It will serve signed gzipped protobufs over http. You can also
restrict your query so that it will only serve known vendors so we can populate
the GUI with vendors immediately from the start.

Json is also available if needed.

We can easily add ssl to the server as well.