Started refactoring for market package
Moved some modules around to make way for new market package that will handle
the (non-kademlia) communication between nodes. Assuming we choose to retain
the existing message format, it's probably easiest to move the rpcudp module
out of the dht package and use it for the market protocol as well.