Add get_metadata message
Adds a new message which returns a short version of the profile. In some
places in the UI we list vendor stores (name, handle, avatar) and need that list
to load as fast as possible. It makes sense to only downloaded a shortened
version of the profile here and only download the full profile when we need it.