Try hole punching if we fail to reach a node.
If an outgoing message fails to reach a node because it's behind a restricted
NAT, we will send a HOLE_PUNCH command to a seed node which will (assuming it's
connected to node behind the NAT) relay the hole_punch message telling that node
to ping us. Once that is done, the original message will be resent and should
get through.