Send raw udp packet in response to hole punch request
Previously when we would receive a hole punch request we would try to send an
empty txrudp packet, which would attempt to establish a new txrudp connection
with the remote peer. Since both nodes on each side were attempting to establish
a new connection, this may have been a source of some of our connection problems.

I switched it to send a raw udp packet to the recipient so that it doesn't attempt
to create a new connection. The packet should be enough to get the NAT to create
a mapping and allow the remote peer to establish a connection like normal.

Fingers crossed this takes care of most of hole punch issues.