Database refactor to store additional moderator info
This commit is a major refactor to the database. It adds an additional field
to the profile for moderation fee (which is also included in the contract) and
changes the moderator table.

Since this changes the db structure for the moderator table, it's a good time
to do the rest of the refactoring to change some primary keys to TEXT. This
will allow better indexing on the tables as well as make it easier to debug
some tables.

Bumps the protocol version because of the profile change.